requirements.txt
Flask==2.0.2
requests==2.26.0
google-generativeai